[Privacy] moves account section from core plugin (#2489)
* moving privacy account core=> main->privacy

* translation privacy and menu jauge

* core/privacy

* privacy core branch translations

* translate desktop -> administration

* delete therm_of_service translation json

* changement domain

* delete core privacy

* correction suite echec PHPUnit
