+++ src/main/app/Resources/modules/layout/sections/administration/components/menu.jsx
@@ -107,9 +107,9 @@ class PlatformStatus extends Component {
             }
           </h3>
 
-          {this.props.disabled && 'La plateforme est désactivée et n\'est plus accessible.'}
-          {!this.props.disabled && !this.props.maintenance && 'La plateforme est ouverte aux utilisateurs.'}
-          {!this.props.disabled && this.props.maintenance && 'La plateforme est fermée et uniquement accessible aux administrateurs.'}
+          {this.props.disabled && trans('platform_opened', {}, 'administration')}
+          {!this.props.disabled && !this.props.maintenance && trans('platform_desactive', {}, 'administration')}
+          {!this.props.disabled && this.props.maintenance && trans('platform_active_admin', {}, 'administration')}
         </div>
       </section>
     )
