+++ src/main/core/Resources/modules/administration/parameters/components/menu.jsx
@@ -25,12 +25,6 @@ const ParametersMenu = (props) =>
           target: props.path+'/',
           exact: true
         }, {
-          name: 'privacy',
-          type: LINK_BUTTON,
-          icon: 'fa fa-fw fa-shield',
-          label: trans('privacy'),
-          target: props.path+'/privacy'
-        }, {
           name: 'appearance',
           type: LINK_BUTTON,
           icon: 'fa fa-fw fa-drafting-compass',
