+++ src/main/core/Resources/translations/platform.en.json
@@ -608,11 +608,6 @@
     "tag": "Tag",
     "tags": "Tags",
     "technical": "Technical",
-    "terms_of_service": "Terms of service",
-    "terms_of_service_activation_message": "Activate the terms of service",
-    "terms_of_service_activation_help": "New users will need to accept the terms to be able to use the platform.",
-    "terms_enabled_but_empty": "Terms of service can only be enabled if a text is provided in at least one language.",
-    "terms_of_service_acceptance": "I accept the terms of service",
     "text": "Text",
     "text_left_align": "Align left",
     "text_length": "Length of the text",
