+++ src/main/core/Resources/translations/platform.fr.json
@@ -626,11 +626,6 @@
     "tag": "Tag",
     "tags": "Tags",
     "technical": "Technique",
-    "terms_of_service": "Conditions d'utilisation",
-    "terms_of_service_activation_message": "Activer les conditions d'utilisation",
-    "terms_of_service_activation_help": "Les nouveaux utilisateurs devront accepter les conditions avant de pouvoir utiliser la plateforme.",
-    "terms_enabled_but_empty": "Les conditions d'utilisation ne peuvent être activées que si le texte est fourni dans au moins une langue.",
-    "terms_of_service_acceptance": "J'accepte les conditions d'utilisation",
     "text": "Texte",
     "text_left_align": "Aligner à gauche",
     "text_length": "Longueur du texte",
