+++ src/main/privacy/Resources/modules/account/privacy/containers/main.jsx
@@ -3,8 +3,8 @@ import {connect} from 'react-redux'
 import {selectors as configSelectors} from '#/main/app/config/store'
 import {selectors as securitySelectors} from '#/main/app/security/store'
 
-import {actions} from '#/main/core/account/privacy/store'
-import {PrivacyMain as PrivacyMainComponent} from '#/main/core/account/privacy/components/main'
+import {actions} from '#/main/privacy/account/privacy/store'
+import {PrivacyMain as PrivacyMainComponent} from '#/main/privacy/account/privacy/components/main'
 
 const PrivacyMain = connect(
   (state) => ({
