+++ src/main/privacy/Resources/modules/administration/privacy/components/tool.jsx
@@ -29,7 +29,7 @@ const PrivacyTool = (props) =>
           fields: [
             {
               name: 'privacy.countryStorage',
-              label: trans('Pays de stockage des données'),
+              label: trans('country_storage', {}, 'privacy'),
               type: 'country'
             }
           ]
@@ -57,18 +57,18 @@ const PrivacyTool = (props) =>
           ]
         }, {
           icon: 'fa fa-fw fa-copyright',
-          title: trans('terms_of_service'),
+          title: trans('terms_of_service', {}, 'privacy'),
           fields: [
             {
               name: 'tos.enabled',
               type: 'boolean',
-              label: trans('terms_of_service_activation_message'),
-              help: trans('terms_of_service_activation_help'),
+              label: trans('terms_of_service_activation_message', {}, 'privacy'),
+              help: trans('terms_of_service_activation_help', {}, 'privacy'),
               linked: [
                 {
                   name: 'tos.text',
                   type: 'translated',
-                  label: trans('terms_of_service'),
+                  label: trans('terms_of_service', {}, 'privacy'),
                   required: true,
                   displayed: get(props.parameters, 'tos.enabled')
                 }
